requirements.txt
accelerate==0.29.3
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
av==14.2.0
bitsandbytes==0.41.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
datasets==2.16.1
DCNv3==1.0
decord==0.6.0
deepspeed==0.14.4
dill==0.3.7
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
flash_attn-2.7.3
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2023.10.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
gradio_client==0.2.9
h11==0.14.0
hf_transfer==0.1.9
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.29.3
idna==3.10
jieba==0.42.1
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
latex2mathml==3.77.0
llava==1.7.0.dev0
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mmcv==2.2.0
mmengine==0.10.7
mmsegmentation==1.2.2
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.15
networkx==3.4.2
ninja==1.11.1.4
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.31.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
peft==0.4.0
pillow==11.1.0
platformdirs==4.3.7
prettytable==3.16.0
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyarrow-hotfix==0.6
pycocotools==2.0.8
pydantic==1.10.8
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.2.2
scipy==1.15.2
sentencepiece==0.1.99
sentry-sdk==2.24.1
setproctitle==1.3.5
shapely==2.0.7
shortuuid==1.0.13
shtab==1.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
svgwrite==1.4.3
sympy==1.13.3
tabulate==0.9.0
termcolor==2.5.0
terminaltables==3.1.10
textrank4zh==0.3
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.20.3
tomli==2.2.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.46.0
triton==2.1.0
typeguard==4.4.2
typing_extensions==4.13.0
tyro==0.9.18
tzdata==2025.2
unrar==0.4
urllib3==1.26.20
uvicorn==0.34.0
wandb==0.18.7
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
